requirements.txt
openai-agents
pytz
requests
python-dotenv
playwright
beautifulsoup4
lxml
pydantic
